server/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart

# LangChain & Ecosystem
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.3.0
langchain-groq>=0.0.1

Vectorstore
pinecone
langchain-pinecone

Embeddings
sentence-transformers
# langchain-google-generative-ai
langchain-google-genai

# PDF Parsing
PyPDF  # backend for PyPDFLoader via langchain_community

# Environment Variables
python-dotenv

# Typing & Utilities
pydantic
requests

# Logging (optional but recommended)
loguru

# File Uploads
aiofiles

# Process
tqdm



# pip install -r requirements.txt